% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disperse}
\alias{disperse}
\title{Seed dispersal across a spatial grid}
\usage{
disperse(S, N, reflect = TRUE, rand = TRUE, seed = 1L)
}
\arguments{
\item{S}{A matrix of seed counts across a spatial grid.}

\item{N}{A neighbor matrix, e.g. produced by \code{neighborhood()}.}

\item{reflect}{Should dispersers exit the domain (\code{FALSE}) or bounce off the domain boundary (\code{TRUE}, default)?}

\item{rand}{Randomize dispersal? (default = \code{TRUE})}

\item{seed}{Integer to seed random number generator.}
}
\value{
A matrix of post-dispersal seed counts of the same dimension as \code{S}.
}
\description{
Seed dispersal across a spatial grid
}
