% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{frame}
\alias{frame}
\title{Convert spatiotemporal population array to tibble}
\usage{
frame(x)
}
\arguments{
\item{x}{Population array generated by \code{simulate}.}
}
\value{
A data frame.
}
\description{
Convert spatiotemporal population array to tibble
}
