% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_heatmaps}
\alias{plot_heatmaps}
\title{Plot population maps for a series of time steps.}
\usage{
plot_heatmaps(d, n_frames = 10)
}
\arguments{
\item{d}{Population array generated by \code{simulate}.}

\item{n_frames}{Number of evenly spaced time steps to plot.}
}
\value{
A ggplot.
}
\description{
Plot population maps for a series of time steps.
}
