% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_lines}
\alias{plot_lines}
\title{Plot population time series in random grid cells}
\usage{
plot_lines(d, n_cells = 10)
}
\arguments{
\item{d}{Population array generated by \code{simulate}.}

\item{n_cells}{Number of grid cells to plot.}
}
\value{
A ggplot.
}
\description{
Plot population time series in random grid cells
}
