% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reproduce}
\alias{reproduce}
\title{Reproduction across a spatial grid}
\usage{
reproduce(N, f)
}
\arguments{
\item{N}{A 3-D array of population numbers for each life stage, over a spatial grid (x, y, class).}

\item{f}{Integer vector of fecundity with a value for each class in \code{N}.}
}
\value{
A matrix.
}
\description{
Reproduction across a spatial grid
}
