% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim}
\alias{sim}
\title{Run a range simulation}
\usage{
sim(
  N,
  env,
  alpha,
  beta,
  gamma,
  fecundity,
  nb,
  reflect = TRUE,
  rand = TRUE,
  seed = 1L,
  record = 0L,
  nsteps = 100L
)
}
\arguments{
\item{N}{A 3-D array of population numbers for each life stage, over a spatial grid (x, y, class).}

\item{env}{A list of environmental data. Each element list element should be a 3-D array of dimensions (x, y, variable).
The list should have one element for time-invariant environment, or \code{nsteps} elements for time-varying environment.}

\item{alpha, }{\code{beta, gamma, fecundity} Demographic parameters; see \code{?transition} and \code{?reproduce}.}

\item{nb}{Neighborhood matrix; e.g. output from \code{neighborhood}.}

\item{rand}{Randomize transitions instead of using matrix multiplication? (Boolean, default = TRUE).}

\item{seed}{Integer to seed random number generator.}
}
\value{
A 3-D array of population numbers for each life stage.
}
\description{
Run a range simulation
}
