% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate}
\alias{simulate}
\title{Run a multi-year range simulation}
\usage{
simulate(sp, ls, diameter = 7, n_steps = 100, record = 3, randomize = TRUE)
}
\arguments{
\item{sp}{Species parameter list, following \code{species_template()}.}

\item{ls}{Landscape spatial data list, following \code{landscape_template()}.}

\item{diameter}{Neighborhood size (integer).}

\item{n_steps}{Number of time steps to simulate (integer).}

\item{record}{Index of age class to record and return (integer).}

\item{randomize}{Should demography and dispersal be randomized (logical)?}
}
\value{
An array of population values over space and time, for the class specified in \code{record}.
}
\description{
Run a multi-year range simulation
}
