% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{species_template}
\alias{species_template}
\title{Generate data structure for species-level parameters.}
\usage{
species_template(n_env = 1, names = c("s", "j", "a"))
}
\arguments{
\item{n_env}{Number of environmental variables (integer).}

\item{names}{Class names (character vector).}
}
\value{
A list of demographic and dispersal parameter objects.
}
\description{
Generate data structure for species-level parameters.
}
