% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disperse}
\alias{disperse}
\title{Seed dispersal across a spatial grid}
\usage{
disperse(S, N)
}
\arguments{
\item{S}{A matrix of seed counts across a spatial grid.}

\item{N}{A neighbor matrix, e.g. produced by \code{neighborhood()}.}
}
\value{
A matrix of post-dispersal seed counts of the same dimension as \code{S}.
}
\description{
Seed dispersal across a spatial grid
}
