% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{neighborhood}
\alias{neighborhood}
\title{Neighborhood dispersal probability matrix.}
\usage{
neighborhood(kernel, diameter = 7, method = "uniform", res = 11)
}
\arguments{
\item{kernel}{A dispersal kernel list, e.g. \code{species_template()$kernel}.}

\item{diameter}{Neighborhood size, in grid cells (odd integer).}

\item{method}{Either "uniform" (default) or "mixed".}

\item{res}{Resolution for numerical integration (odd integer).}
}
\value{
A matrix of dispersal probabilities.
}
\description{
This function uses numerical integration of a dispersal kernel to construct
a matrix representing the probability that dispersal originating in a given
grid cell will arrive at each cell across its neighborhood.
}
