% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{neighborhood}
\alias{neighborhood}
\title{Neighborhood dispersal probability matrix}
\usage{
neighborhood(
  kernel,
  diameter = 7,
  cell_res,
  method = c("area", "area-centroid", "centroid"),
  res = 11
)
}
\arguments{
\item{kernel}{A dispersal kernel object, e.g. generated by \code{species_template()$kernel}.
Note that the kernel function must already be area-adjusted by incorporating a denominator of 2*pi,
as has been done for the included functions \code{dlognormal}, \code{d2Dt}, and \code{dexponential}.
Distance units are assumed to be in meters.}

\item{diameter}{Neighborhood size, in grid cells (odd integer).}

\item{cell_res}{Grid cell size, in meters.}

\item{method}{Either "area" (default), "area-centroid", or "centroid"; see Chipperfield et al. (2011).}

\item{res}{Resolution for numerical integration (odd integer).}
}
\value{
A matrix of dispersal probabilities.
}
\description{
This function uses numerical integration of a dispersal kernel to construct
a matrix representing the probability that dispersal originating in a given
grid cell will arrive at each cell across its neighborhood.
}
