% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate}
\alias{simulate}
\title{Run a range simulation}
\usage{
simulate(
  sp,
  ls,
  n_steps = 100,
  randomize = TRUE,
  reflect = TRUE,
  record = 3,
  ...
)
}
\arguments{
\item{sp}{Species parameter list, following \code{species_template()}.}

\item{ls}{Landscape spatial data list, following \code{landscape_template()}.}

\item{n_steps}{Number of time steps to simulate (integer).}

\item{randomize}{Should demography and dispersal be randomized (logical)?}

\item{reflect}{Should dispersers bounce off domain boundary (logical)?}

\item{record}{Index of age class to record and return (integer).}

\item{...}{Further arguments passed to \code{neighborhood()}.}
}
\value{
An array of population values over space and time, for the class specified in \code{record}.
}
\description{
Run a range simulation
}
