% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transition}
\alias{transition}
\title{Perform a randomized stage-based demographic transition}
\usage{
transition(N, E, alpha, beta, gamma, rand = TRUE, seed = 1L)
}
\arguments{
\item{N}{A 3-D array of population numbers for each life stage, over a spatial grid (x, y, class).}

\item{E}{A 4-D array of environmental data (x, y, time, variable).}

\item{alpha}{A matrix of transition intercepts (to, from).}

\item{beta}{A 3-D array of density dependence effects (to, from, modifier).}

\item{gamma}{A 3-D array of environmental effects (to, from, variable).}

\item{rand}{Randomize transitions instead of using matrix multiplication? (Boolean, default = TRUE).}

\item{seed}{Integer to seed random number generator.}
}
\value{
A 3-D array of population numbers for each life stage.
}
\description{
Perform a randomized stage-based demographic transition
}
